pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-factory"
version = "0.1.0"
description = "Vision-to-JSON Pipeline for Database Creation"
readme = "README.md"
authors = [
    { name = "Advait Pujari" }
]
requires-python = ">=3.9"
dependencies = [
    "boto3==1.34.40",
    "pdf2image==1.17.0",
    "Pillow==10.2.0",
    "pydantic==2.6.1",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "tenacity==8.2.3"
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "black==24.1.1",
    "flake8==7.0.0"
]

[project.scripts]
vision-factory = "vision_factory.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vision_factory*"]
